requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
attrdict==2.0.1
attrs==25.4.0
audioread==3.1.0
bitsandbytes==0.45.3
braceexpand==0.1.7
brotli==1.2.0
cdifflib==1.2.6
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
cytoolz==1.1.0
datasets==4.4.1
decorator==5.2.1
dill==0.4.0
distance==0.1.3
docopt==0.6.2
editdistance==0.8.1
einops==0.8.1
executing==2.2.1
fastapi==0.121.1
ffmpy==0.6.4
fiddle==0.3.0
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2024.12.0
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.12
gitpython==3.1.45
gradio==5.49.1
gradio-client==1.13.3
graphviz==0.21
greenlet==3.2.4
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.1.2
hydra-core==1.3.2
idna==3.11
inflect==7.5.0
intervaltree==3.1.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
janome==0.5.0
jedi==0.19.2
jieba==0.42.1
jinja2==3.1.6
jiwer==3.1.0
joblib==1.5.2
kaldi-python-io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.9
kornia==0.8.2
kornia-rs==0.1.10
lazy-loader==0.4
levenshtein==0.27.3
lhotse==1.31.1
libcst==1.8.6
librosa==0.11.0
lightning==2.4.0
lightning-utilities==0.15.2
lilcom==1.8.1
llvmlite==0.44.0
loguru==0.7.3
mako==1.3.10
markdown==3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==4.1.0
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.1.6
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
nemo-text-processing==1.1.0
nemo-toolkit==2.3.0
networkx==3.5
nltk==3.9.2
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
onnx==1.17.0
optuna==4.6.0
orjson==3.11.4
packaging==24.2
pandas==2.3.3
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==11.3.0
plac==1.4.5
platformdirs==4.5.0
pooch==1.8.2
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==4.24.4
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
pyannote-core==5.0.0
pyannote-database==5.1.3
pyannote-metrics==3.2.1
pyarrow==22.0.0
pybind11==3.0.1
pycparser==2.23
pydantic==2.11.10
pydantic-core==2.33.2
pydub==0.25.1
pygments==2.19.2
pyloudnorm==0.1.1
pynini==2.1.6.post1
pyparsing==3.2.5
pypinyin==0.55.0
pypinyin-dict==0.9.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.6
pytz==2025.2
pyyaml==6.0.3
rapidfuzz==3.14.3
regex==2025.11.3
requests==2.32.5
resampy==0.4.3
rich==14.2.0
ruamel-yaml==0.18.16
ruamel-yaml-clib==0.2.14
ruff==0.14.4
sacremoses==0.1.1
safehttpx==0.1.7
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.43.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
sqlalchemy==2.0.44
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.13.3
toolz==1.1.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchmetrics==1.8.2
torchvision==0.22.1+cu118
tqdm==4.67.1
traitlets==5.14.3
transformers @ git+https://github.com/huggingface/transformers.git@756742354b5cde1afadeba584a6c0e1577bf0148
triton==3.3.1
typeguard==4.4.4
typer==0.20.0
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wandb==0.22.3
wcwidth==0.2.14
webdataset==1.0.2
websockets==15.0.1
werkzeug==3.1.3
wget==3.2
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
